# docs/api/library-v1.openapi.yaml
openapi: 3.1.0
...
info:
  title: XiaDown Library API
  version: 1.0.0
  description: |
    Version 1 of the device-facing XiaDown Library API. The API is exposed only
    by a user-enabled Library transport. Pairing credentials and Bearer tokens
    are secrets; clients must not put them in URLs, analytics, logs, or crash
    reports.

    The pairing response grants `library.read` and `tasks.read` by default.
    `music.read`, `music.state`, `music.manage`, `rss.read`, `rss.state`,
    `rss.manage`, `rss.fetch`, `tasks.create`, and `tasks.control` can be granted
    later by the desktop owner. Pairing never grants either RSS management or
    public-feed fetch access by default.
    A grant is bound to one Catalog, so a valid token from another Catalog is
    rejected.

    A stable device ID and public-key hash are client assertions, not authority
    to inherit an existing grant. Pairing rejects collisions with an active
    grant. After the Desktop owner revokes a grant, a fresh short-lived pairing
    session may replace it in place: the grant identity and creation time remain
    for namespace continuity, but the token and public-key hash rotate and scopes
    reset to the default minimum. The old token remains invalid and prior managed
    scopes must be granted again by the Desktop owner.

    Canonical response property names are case-sensitive camelCase. Path and
    query parameter names are case-sensitive. Unless an input is explicitly
...
servers:
...
    variables:
...
security: []
...
paths:
...
  /api/v1/sync/events:
    get:
      tags: [Sync]
      operationId: streamSyncEvents
      summary: Receive scope-filtered synchronization wake hints
      description: |-
        Foreground-only Server-Sent Events signal plane, advertised by the
        top-level `sync-events-v1` capability in `GET /api/v1/device-access`.
        `station-dirty` carries only a Station name, epoch, and high-water mark;
        `tasks-dirty` carries only `{ "dirty": true }`. Clients must use the
        corresponding overview/changes or task-list endpoint after every hint.
        Event loss, disconnect, suspension, replay gaps, and a missing stream do
        not affect correctness. `Last-Event-ID` only reduces duplicate wakeups
        and is never a Station cursor or acknowledgement. Desktop periodically
        revalidates the presented Bearer grant while the stream is active.
        Revocation or expiry closes the stream without an SSE error payload;
        scope changes atomically update filtering and discard any pending hint
        that is no longer authorized before it can be delivered.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/LastEventID"
      responses:
        "200":
          description: |-
            An SSE stream containing `station-dirty`, `tasks-dirty`, and
            periodic comment keepalives. Events are filtered by the current
            grant's `library.read`, `music.read`, `rss.read`, and `tasks.read`
            scopes. Slow consumers receive the newest coalesced hint per
            Station rather than an unbounded queue.
          content:
            text/event-stream:
              schema:
                type: string
              examples:
                stationDirty:
                  value: |-
                    id: 907
                    event: station-dirty
                    data: {"station":"music","epoch":"0123456789abcdef0123456789abcdef","highWater":1881}
        "400":
          description: Invalid numeric `Last-Event-ID` (`invalid_last_event_id`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: The credential belongs to a different Catalog (`catalog_access_denied`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: The HTTP transport cannot stream responses (`streaming_unavailable`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: xd1 opaque device token
      description: "Send `Authorization: Bearer <token>`. The token is Catalog-bound and must be stored as a secret."
...
  parameters:
    LastEventID:
      name: Last-Event-ID
      in: header
      description: |-
        Optional decimal SSE event ID from the last delivered hint. It is a
        best-effort replay optimization only and must not be stored or compared
        as a Library, Music, or RSS synchronization cursor.
      schema:
        type: string
        maxLength: 20
        pattern: '^[1-9][0-9]*$'
...
  responses:
...
    Unauthorized:
      description: Missing, malformed, expired, revoked, or invalid Bearer token (`unauthorized`).
      headers:
        WWW-Authenticate:
          description: Bearer challenge for the XiaDown Library realm.
          schema:
            type: string
            const: 'Bearer realm="xiadown-library"'
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  schemas:
    ErrorCode:
      type: string
      description: Stable machine-readable API error code. Error messages and implementation details are intentionally absent.
      enum:
        - asset_stream_limit_reached
        - asset_unavailable
        - catalog_access_denied
        - insufficient_scope
        - idempotency_conflict
        - internal_error
        - invalid_cursor
        - invalid_music_mutation
        - invalid_pagination
        - invalid_request
        - music_content_changed
        - music_dependency_pending
        - music_entity_not_found
        - music_idempotency_conflict
        - music_revision_conflict
        - music_stream_limit_reached
        - not_found
        - pairing_failed
        - pairing_invalid
        - reset_required
        - rss_state_conflict
        - task_state_conflict
        - unauthorized
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ErrorCode"
...